add different tests at jsdom & happydom
