+++ package.json
@@ -25,7 +25,6 @@
   "devDependencies": {
     "@cspell/eslint-plugin": "^9.6.0",
     "@eslint/js": "^9.39.2",
-    "@playwright/test": "^1.59.1",
     "@testing-library/dom": "^10.4.1",
     "@types/node": "^25.6.0",
     "@vitest/coverage-v8": "^4.1.2",
